### 2021-06-26 #3420 @whedon
[PRE REVIEW]: daiR: an R package for OCR with Google Document AI
**Submitting author:** @Hegghammer (<a href="http://orcid.org/0000-0001-6253-1518">Thomas Hegghammer</a>)
**Repository:** <a href="https://github.com/Hegghammer/daiR" target ="_blank">https://github.com/Hegghammer/daiR</a>
**Version:** v.0.9.1
