### 2021-06-26 #3420 @arfon
@Hegghammer – thanks for your submission to JOSS. It looks like the YAML header of your file isn't formatted quite right, please take a look at the example and modify accordingly: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

Also, please take a look at the DOI suggestions from Whedon and update your BibTeX file if they look correct.

Finally, we're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
