### 2021-06-27 e922a8d
fix yaml and DOIs
