+++ paper.bib
@@ -34,6 +34,7 @@
 	journal = {Socarxiv},
 	year = 2021,
 	month = Jun,
+	doi = {10.31235/osf.io/6zfvs},
 	url = {https://osf.io/preprints/socarxiv/6zfvs},
 }
 
@@ -58,6 +59,7 @@
   author = {{Santos}, E.A.},
   booktitle = {Proceedings of the Workshop on Computational Methods for Endangered Languages},
   volume = 1,
+  doi = {10.33011/computel.v1i.345},
   year = 2019,
 }
 
