### 2021-06-27 #3420 @Hegghammer
Dear Arfon Smith,
Many thanks for the quick response. I have addressed the YAML and DOI
issues, and updated versions of paper.md and paper.bib are now in the
Github repository.
No problem at all with the backlog. JOSS is already lightyears ahead of the
journals I normally deal with in terms of efficiency.
Thanks again and all the best,
Thomas Hegghammer
