### 2021-07-05 1ad0570
handle NAs in box coordinates
