### 2021-07-09 #3420 @Hegghammer
Thank you for considering the manuscript! My suggestions, in order of
perceived topic fit:

   - lmullen
   - josiahparry
   - neuroimaginador
   - adamhsparks

Best,
Thomas

On Fri, 9 Jul 2021 at 10:04, Nikoleta Glynatsi ***@***.***>
wrote:

> @Hegghammer <https://github.com/Hegghammer> thank you again for your
> submission with JOSS, if you have not already done so, could you take a
> look through https://bit.ly/joss-reviewers to suggest some potential
> reviewers 😄
>
