### 2021-07-10 #3420 @Nikoleta-v3
👋 @lmullen & @josiahparry & @adamhsparks & @neuroimaginador, would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "daiR: an R package for OCR with Google Document AI". You can find more information at the top of this Github issue 🆙
