### 2021-07-20 #3420 @Nikoleta-v3
👋🏻 @expectopatronum & @hiendn & @shivam11 would any of you be willing to review this submission for JOSS?

Currently looking for one more reviewer! The submission I'd like you to review is titled: "daiR: an R package for OCR with Google Document AI". You can find more information at the top of this Github issue 🆙
