### 2021-07-25 #3420 @shivam11
Yes, I am available.RegardsShivam    On Tuesday, July 20, 2021, 12:25:34 PM GMT+4, Nikoleta Glynatsi ***@***.***> wrote:

👋🏻 @expectopatronum & @hiendn & @shivam11 would any of you be willing to review this submission for JOSS?

Currently looking for one more reviewer! The submission I'd like you to review is titled: "daiR: an R package for OCR with Google Document AI". You can find more information at the top of this Github issue 🆙
