### 2021-07-26 #3538 @whedon
[REVIEW]: daiR: an R package for OCR with Google Document AI
**Submitting author:** @Hegghammer (<a href="http://orcid.org/0000-0001-6253-1518">Thomas Hegghammer</a>)
**Repository:** <a href="https://github.com/Hegghammer/daiR" target ="_blank">https://github.com/Hegghammer/daiR</a>
**Version:** v0.9.2
