### 2021-08-17 #3538 @cjbarrie
I am almost finished with my review. A few comments follow. The only check box remaining is the community guidelines, which I don't see on the Github Repo. The remainder of my comments pertain to the software paper. This is a great package and it deserves its place in JOSS.

- Acronym DAI should be in parentheses after Google Document AI (line 18) before being used as acronym in line 19.

- Line 28 “together allow for” not allows?

- I wold drop the paragraphs lines 30-38 as they go into detail about current API functionality and package functionality, which may change in near future. In other words, keep the description to the high-level functionality? The same goes for detailing cost per 1000 pages. I think this detail can be described in package documentation rather than a software paper like this.

- Worth adding code of conduct to Github Repo: see e.g. https://github.com/cjbarrie/academictwitteR#code-of-conduct
