+++ paper.md
@@ -13,95 +13,31 @@ authors:
 affiliations:
   - name: Senior Research Fellow, Norwegian Defence Research Establishment (FFI)
     index: 1
-date: 25 June 2021
+date: 31 August 2021
 bibliography: paper.bib
 ---
 
 # Statement of need
 
-Optical character recognition (OCR) promises to open up centuries worth of
-text to computational analysis. But OCR software has long been sensitive to
-visual noise and weak on non-Western languages. In April 2021, Google launched
-Document AI (DAI), a server-based processor offering high-accuracy OCR for
-over sixty languages [@vanguri:2021]. The `daiR` [@hegghammer:2021] package
-provides an R interface to the Document AI API along with additional tools for
-output parsing and visualization.
+Optical character recognition (OCR) promises to open up centuries worth of text to computational analysis. But OCR software has long been sensitive to visual noise and weak on non-Western languages. In April 2021, Google launched Document AI (DAI), a server-based processor offering high-accuracy OCR for over sixty languages [@vanguri:2021]. The `daiR` [@hegghammer:2021] package provides an R interface to the Document AI API along with additional tools for output parsing and visualization.
 
 # Summary
 
-Text as data is a growing field in the social sciences and digital humanities,
-but computational access to text produced before the late 20th century has been 
-limited by the difficulty of extracting text from document scans. Established 
-OCR libraries such as Tesseract [@tesseract:2021] are highly sensitive to noise 
-and often require extensive corpus-specific adaptations to render text accurately.
+Text as data is a growing field in the social sciences and digital humanities, but computational access to text produced before the late 20th century has been limited by the difficulty of extracting text from document scans. Established OCR libraries such as Tesseract [@tesseract:2021] are highly sensitive to noise and often require extensive corpus-specific adaptations to render text accurately.
 
-The past two years have seen the introduction of server-based OCR processors,
-such as Amazon Textract [@amazon:2021] and Google Document AI, which offer
-very high accuracy out of the box [@hegghammer:2021b]. Of the two, DAI performs
-better in benchmarking tests and offers broader language support. 
+The past two years have seen the introduction of server-based OCR processors, such as Amazon Textract [@amazon:2021] and Google Document AI (DAI), which offer very high accuracy out of the box [@hegghammer:2021b]. Of the two, DAI performs better in benchmarking tests and offers broader language support.
 
-In R, where many scholars do their text analysis work, there are packages
-for Tesseract [@ooms:2021] and Amazon Textract [@kretch:2021], but not for
-Document AI. The primary objective of `daiR` is therefore to provide access,
-from within R, to all the main functionalities of the Document AI API. The
-secondary aim is to offer tools to help parse the output returned by
+In R, where many scholars do their text analysis work, there are packages for Tesseract [@ooms:2021] and Amazon Textract [@kretch:2021], but not for Document AI. The primary objective of `daiR` is therefore to provide access, from within R, to all the main functionalities of the Document AI API. The secondary aim is to offer tools to help parse the output returned by
 the DAI processor.
 
-DAI is part of Google Cloud Services (GCS), a suite of cloud computing
-services for storage, analytics, and machine learning. `daiR` joins a family
-of existing R packages that interface with GCS, such as `googleLanguageR`
-[@edmondson:2020], that together allows for the implementation of multiple
-GCS tools into an R-based text mining workflow.
[…]
