### 2021-08-31 #3538 @Hegghammer
Thanks for your understanding. I have now addressed all of @cjbarrie
<https://github.com/cjbarrie>'s comments, pushed the changes to the
repository, and checked that the paper compiles.
Best,
Thomas

On Sat, 28 Aug 2021 at 10:06, Nikoleta Glynatsi ***@***.***>
wrote:

> Not a problem @Hegghammer <https://github.com/Hegghammer>! Thank you for
> the update 😄
>
