### 2021-11-09 #3538 @Nikoleta-v3
👋🏼  @geraintpalmer  this is the review thread for the paper. Thank you for agreeing to review this submission!

You have a checklists at the top of this thread with the JOSS requirements 🔝 As you go over the submission, please check any items that you feel have been satisfied. There are also links to the JOSS reviewer guidelines.

Please feel free to ping me (@Nikoleta-v3 ) if you have any questions/concerns 👍🏻 😄
