### 2021-12-10 #3538 @Nikoleta-v3
Hey @cjbarrie 😄  Thank you so much for your suggestion and for taking the time to suggest a second reviewer!

I have spoken with @geraintpalmer and I know that he is currently reviewing the paper. So the submission should be moving forward very soon.

I am a new editor and I hadn’t thought of asking you (or someone who has completed their review) for suggestions in case the other reviewer dropped out. That’s something I will be doing in the future so sure. Thank you again! 🙏🏻
