### 2021-12-16 #3538 @geraintpalmer
Nearly done with the review, sorry for the delay.

@Hegghammer which License do you use? It looks a little bare, is it one listed here: https://opensource.org/licenses/alphabetical ?
