### 2021-12-16 #3538 @Hegghammer
Thanks for reviewing! The license is MIT as per the description file, with
the license file serving as a supplement. I'm not a specialist on licenses,
so I don't know if it's ideal, but I've seen several other packages do it
this way.
Best
Thomas

On Thu, 16 Dec 2021 at 21:50, Geraint Palmer ***@***.***>
wrote:

> Nearly done with the review, sorry for the delay.
>
> @Hegghammer <https://github.com/Hegghammer> which License do you use? It
> looks a little bare, is it one listed here:
> https://opensource.org/licenses/alphabetical ?
>
