### 2021-12-16 #3538 @geraintpalmer
Thanks @Hegghammer, I'm not a license expert either, but I usually see the text of the licence copy and pasted into the LICENSE file which makes things much clearer. E.g https://github.com/jquery/jquery/blob/main/LICENSE.txt

Thanks, happy to accept the review then.
