### 2021-12-17 #3538 @Hegghammer
That is a good suggestion which I will gladly incorporate. I will edit the
LICENCE file accordingly.
Many thanks for everyone's time and best wishes for the holidays.
Best,
Thomas

On Fri, 17 Dec 2021 at 00:14, Geraint Palmer ***@***.***>
wrote:

> Thanks @Hegghammer <https://github.com/Hegghammer>, I'm not a license
> expert either, but I usually see the text of the licence copy and pasted
> into the LICENSE file which makes things much clearer. E.g
> https://github.com/jquery/jquery/blob/main/LICENSE.txt
>
> Thanks, happy to accept the review then.
>
