### 2021-12-19 #3538 @Nikoleta-v3
Thank you very much for your patience @Hegghammer! Everything looks good 👍🏻

One minor comment regarding the point:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Could you please add a `Contributing` file to the project's repository?

Here is an example from the last submission I was handling: https://github.com/mscroggs/symfem/blob/main/CONTRIBUTING.md.

After you added the file could you:

 - [ ]  Make a new tagged release of your software (only if necessary, the current tag is ` v.0.9.1` ), and list the version tag of the archived version here.

I can then move forward with accepting the submission
