+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,22 @@
+---
+name: Bug report
+about: Report what you think is a problem in the code
+title: "[BUG]:"
+labels: bug
+assignees: Hegghammer
+
+---
+
+**Describe the problem**
+A concise description of the bug
+
+**To Reproduce**
+Steps to reproduce the behavior
+
+** Your configuration**
+ - OS: [e.g. Windows 10]
+ - R version [e.g. R 4.1.2]
+ - IDE type and version: [e.g. RStudio 2021.09.1]
+
+**Additional context and screenshots**
+Add any other context about the problem here. If applicable, add screenshots to help explain your problem.
