+++ .github/ISSUE_TEMPLATE/technical-problem.md
@@ -1,0 +1,22 @@
+---
+name: Technical problem
+about: Something's not working and you're not sure why
+title: "[PROBLEM]:"
+labels: help wanted
+assignees: Hegghammer
+
+---
+
+**Describe the problem**
+A concise description of the problem
+
+**Summarize the procedure**
+A description of the steps you followed until the problem occurred (the more detailed the better)
+
+** Your configuration**
+ - OS: [e.g. Windows 10]
+ - R version [e.g. R 4.1.2]
+ - IDE type and version: [e.g. RStudio 2021.09.1]
+
+**Additional context and screenshots**
+Add any other context about the problem here. If applicable, add screenshots to help explain your problem.
