### 2021-12-19 615bc5a
add code of conduct + contributing files
