+++ CONTRIBUTING.md
@@ -1,0 +1,18 @@
+## How to contribute
+
+### Reporting bugs
+If you find a bug in daiR, please report it on the [issue tracker](https://github.com/Hegghammer/daiR/issues/new?assignees=Hegghammer&labels=bug&template=bug_report.md&title=%5BBUG%5D%3A).
+
+### Suggesting enhancements
+If you want to suggest a new feature or an improvement of a current feature, you can submit this
+on the [issue tracker](https://github.com/Hegghammer/daiR/issues/new?assignees=Hegghammer&labels=enhancement&template=feature_request.md&title=%5BFEATURE%5D%3A).
+
+### Reporting other problems
+If you encounter a problem using daiR and are not sure what the source of the problem is, you can report this too on the [issue tracker](https://github.com/Hegghammer/daiR/issues/new?assignees=Hegghammer&labels=help+wanted&template=technical-problem.md&title=%5BPROBLEM%5D%3A)
+
+### Submitting a pull request
+If you want to directly submit code to daiR, you can do this by forking the daiR repo and then submitting a pull request.
+
+### Code of conduct
+We expect all our contributors to follow [the Contributor Covenant](CODE_OF_CONDUCT.md). Any unacceptable
+behaviour can be reported to Thomas (hegghammer@gmail.com).
