### 2021-12-19 a7441be
add JOSS reviewers to acknowledgments
