### 2021-12-19 #3538 @Hegghammer
Dear Nikoleta,

Many thanks for your message. The sample Contributing file was very useful.

I have now:
- Added a CONTRIBUTING.md file and created issue templates
- Added a CODE_OF_CONDUCT.md file
- Made a new tagged release (
https://github.com/Hegghammer/daiR/releases/tag/v0.9.2)
- Archived the 0.9.2 version of the repo on Zenodo (
https://zenodo.org/record/5792037) under the DOI 10.5281/zenodo.5792037.

I checked the Zenodo metadata to the best of my ability, but wasn't
entirely sure how you like the title etc to be formatted, so just let me
know if I should change anything.

Thanks again for considering my submission. I remain at your disposal for
any further requests or questions.

Best,
Thomas

On Sun, 19 Dec 2021 at 13:08, Nikoleta Glynatsi ***@***.***>
wrote:

> Thank you very much for your patience @Hegghammer
> <https://github.com/Hegghammer>! Everything looks good 👍🏻
>
> One minor comment regarding the point:
>
> Community guidelines: Are there clear guidelines for third parties wishing
> to 1) Contribute to the software 2) Report issues or problems with the
> software 3) Seek support
>
> Could you please add a Contributing file to the project's repository?
>
> Here is an example from the last submission I was handling:
> https://github.com/mscroggs/symfem/blob/main/CONTRIBUTING.md.
>
> After you added the file could you:
>
>    - Make a new tagged release of your software (only if necessary, the
>    current tag is v.0.9.1 ), and list the version tag of the archived
>    version here.
>    - Archive the reviewed software in Zenodo or a similar service (e.g.,
>    figshare, an institutional repository)
>    - Check the archival deposit (e.g., in Zenodo) has the correct
>    metadata. This includes the title (should match the paper title) and author
>    list (make sure the list is correct and people who only made a small fix
>    are not on it). You may also add the authors' ORCID.
>    - Please list the DOI of the archived version here.
>
> I can then move forward with accepting the submission
>
