### 2021-12-20 #3538 @Nikoleta-v3
Thank you @Hegghammer ! The `CONTRIBUTING.md` & `CODE_OF_CONDUCT.md` look great 😄
